Refactor derive_sertype in sertype interface: include data representation and type consistency enforcement qos parameters
